requirements.txt
Flask==3.0.3
gspread==6.0.0
google-auth==2.29.0
gunicorn==21.2.0